`refactor(dao): migrate BookDao to DAO pattern with new BookDAO interface and BookDAOImpl implementation`
